various minor bits: add miscellaneous comment fixes; in get_next_filename(), use a long instead of an int for the number prepended to the filename; and in num_of_digits(), use a ssize_t instead of an int
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@2468 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/files.c
@@ -379,11 +379,11 @@ int open_file(const char *filename, bool newfie, FILE **f)
  * extension exists, we return "". */
 char *get_next_filename(const char *name)
 {
-    int i = 0;
+    long i = 0;
     char *buf;
     size_t namelen = strlen(name);
 
-    buf = charalloc(namelen + num_of_digits(INT_MAX) + 7);
+    buf = charalloc(namelen + num_of_digits(LONG_MAX) + 7);
     strcpy(buf, name);
     strcpy(buf + namelen, ".save");
     namelen += 5;
@@ -393,15 +393,16 @@ char *get_next_filename(const char *name)
 
 	if (stat(buf, &fs) == -1)
 	    return buf;
-	if (i == INT_MAX)
+	if (i == LONG_MAX)
 	    break;
 
 	i++;
-	sprintf(buf + namelen, ".%d", i);
+	sprintf(buf + namelen, ".%ld", i);
     }
 
     /* We get here only if there is no possible save file. */
     null_at(&buf, 0);
+
     return buf;
 }
