+++ src/utils.c
@@ -55,7 +55,7 @@ int regexp_bol_or_eol(const regex_t *preg, const char *string)
 }
 #endif /* HAVE_REGEX_H */
 
-int num_of_digits(int n)
+int num_of_digits(ssize_t n)
 {
     int i = 1;
